feat: add optional `clone` method to `FormatPlugin` interface
This method is used to clone the object returned by `read` before
passing it to `update`.

This is useful when the object returned by `read` is not trivially
cloneable. One use case is when trying to preserve comments in a JSONC
file using the `comment-json` package, which uses symbol properties on
the object to store comments. The default `clone` method would not
preserve these symbol properties, so a custom `clone` method is needed.

Fixes #10